requirements.txt
pymysql
pillow
django==4.0.0
django-taggit
matplotlib

# celery
celery
django-celery-beat
django-celery-results
flower

# for api
djangorestframework
django-filter

# only admin
django-simpleui
django-import-export
django-admin-model-list-order
django-admin-sortable
django-prettyjson
django-cascading-dropdown-widget
django-attachments

# test library
requests
